+++ Changes.md
@@ -1,3 +1,4 @@
+* #973 - Show a reasonable error message when a duplicate performance score is entered
 * #971, #972 - make review mode button show as depressed when review mode is on
 * #1024 - Allow judges to see comments on the score entry page
 * #1031 - add support for assigning tables for playoffs without printing